Refactor contributors page
