+++ src/features/organizer/Organizer.jsx
@@ -2,7 +2,8 @@ import React, { memo } from 'react'
 import { Route, Routes } from 'react-router-dom'
 
 import AppLayout from 'layout'
-import Profile from 'features/profile/profile'
+import Profile from 'features/profile'
+import Contributors from 'features/contributors'
 import Sidebar from './sidebar'
 
 const Organizer = () => {
@@ -10,6 +11,7 @@ const Organizer = () => {
     <AppLayout sidebar={<Sidebar />}>
       <Routes>
         <Route path="profile" element={<Profile />} />
+        <Route path="contributors" element={<Contributors />} />
       </Routes>
     </AppLayout>
   )
